Rename app_apply_for_membership -> app_membership
Need to knife database in prod for this to work: https://stackoverflow.com/questions/8408046/how-to-change-the-name-of-a-django-app
