--- prodekoorg/app_apply_for_membership/__init__.py
@@ -1,3 +1,0 @@
-default_app_config = (
-    "prodekoorg.app_apply_for_membership.apps.AppApplyForMembershipConfig"
-)
