+++ prodekoorg/app_membership/__init__.py
@@ -1,0 +1,3 @@
+default_app_config = (
+    "prodekoorg.app_membership.apps.AppApplyForMembershipConfig"
+)
