+++ prodekoorg/app_membership/apps.py
@@ -3,8 +3,8 @@ from django.utils.translation import ugettext_lazy as _
 
 
 class AppApplyForMembershipConfig(AppConfig):
-    name = "prodekoorg.app_apply_for_membership"
+    name = "prodekoorg.app_membership"
     verbose_name = _("Membership applications")
 
     def ready(self):
-        import prodekoorg.app_apply_for_membership.signals
+        import prodekoorg.app_membership.signals
