+++ prodekoorg/app_membership/templates/app_membership_base.html
@@ -3,9 +3,9 @@
 
 {% block extra_head %}
     {% if not done and not error %}
-        <script src="{% static 'js/apply_for_membership-custom.js' %}"></script>
+        <script src="{% static 'js/app_membership-custom.js' %}"></script>
     {% endif %}
-    <link href="{% sass_src 'scss/app_apply_for_membership.scss' %}" rel="stylesheet" type="text/css" />
+    <link href="{% sass_src 'scss/app_membership.scss' %}" rel="stylesheet" type="text/css" />
 {% endblock %}
 
 {% block content %}
