+++ prodekoorg/app_membership/templates/emails/accept_mail_fi.txt
