+++ prodekoorg/app_membership/templates/emails/info_mail.html
@@ -10,4 +10,4 @@
 <p>Muuta: {{ user.additional_info }}</p>
 <p>{% endif %}</p>
 <p>&nbsp;</p>
-<p>https://www.prodeko.org/fi/admin/app_apply_for_membership/pendinguser/{{ user.id }}/change/ - Hyv&auml;ksy tai hylk&auml;&auml; hakemus t&auml;st&auml;. Hyväksytylle hakijalle l&auml;hetet&auml;&auml;n s&auml;hk&ouml;postilla k&auml;ytt&auml;j&auml;tunnus ja salasana prodeko.orgiin. Sama tunnus käy matrikkeliin.</p>
+<p>https://www.prodeko.org/fi/admin/app_membership/pendinguser/{{ user.id }}/change/ - Hyv&auml;ksy tai hylk&auml;&auml; hakemus t&auml;st&auml;. Hyväksytylle hakijalle l&auml;hetet&auml;&auml;n s&auml;hk&ouml;postilla k&auml;ytt&auml;j&auml;tunnus ja salasana prodeko.orgiin. Sama tunnus käy matrikkeliin.</p>
