+++ prodekoorg/app_utils/signals.py
@@ -4,7 +4,7 @@ from django.dispatch import receiver
 from easy_thumbnails.signal_handlers import generate_aliases_global
 from easy_thumbnails.signals import saved_file
 from lifelonglearning.models import Course
-from prodekoorg.app_apply_for_membership.models import PendingUser
+from prodekoorg.app_membership.models import PendingUser
 from prodekoorg.app_infoscreen.models import Slide
 from prodekoorg.app_kulukorvaus.models import Kulukorvaus, KulukorvausPerustiedot
 from prodekoorg.app_poytakirjat.models import Dokumentti
