+++ prodekoorg/templates/layout-components/footer.html
@@ -23,7 +23,7 @@
                 <div class="col-xl-7 col-sm-12 col-7 pl-xl-2">
                     <p><u><a href="{% url 'app_contact:contact' %}">{% trans contact_form %}</a></u></p>
                     <p><u><a href="{% url 'app_kulukorvaus:kulukorvaus' %}">{% trans reimbursements %}</a></u></p>
-                    <p><u><a href="{% url 'app_apply_for_membership:apply' %}">{% trans apply_for_membership %}</a></u></p>
+                    <p><u><a href="{% url 'app_membership:apply' %}">{% trans apply_for_membership %}</a></u></p>
                     <p><u><a href="/privacy-policy">{% trans privacy_policy %}</a></u></p>
                 </div>
             </div>
